+++ CHANGELOG.md
@@ -56,6 +56,9 @@ END_UNRELEASED_TEMPLATE
 ### Removed
 * (coverage) Support for python 3.8 has been dropped from the bundled
   `coverage.py` wheel set, since coverage.py 7.6.2 dropped it.
+* (workspace) Users loading `rules_python` through `WORKSPACE` will have
+  to migrate to loading it through `MODULE.bazel`. The previous release is
+  the last release that supports `WORKSPACE`.
 
 {#v0-0-0-changed}
 ### Changed